male modifikacije

+++ otc/src/main/java/rs/edu/raf/si/bank2/otc/services/CompanyBankAccountService.java
@@ -5,8 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rs.edu.raf.si.bank2.otc.dto.CompanyBankAccountDto;
 import rs.edu.raf.si.bank2.otc.exceptions.BankAccountNotFoundException;
+import rs.edu.raf.si.bank2.otc.exceptions.CompanyNotFoundException;
+import rs.edu.raf.si.bank2.otc.models.mongodb.Company;
 import rs.edu.raf.si.bank2.otc.models.mongodb.CompanyBankAccount;
 import rs.edu.raf.si.bank2.otc.repositories.mongodb.CompanyBankAccountRepository;
+import rs.edu.raf.si.bank2.otc.repositories.mongodb.CompanyRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,6 +19,9 @@ public class CompanyBankAccountService {
 
     @Autowired
     private CompanyBankAccountRepository bankAccountRepository;
+
+    @Autowired
+    private CompanyRepository companyRepository;
 
     public List<CompanyBankAccount> getAllBankAccounts() {
         return bankAccountRepository.findAll();
@@ -30,8 +36,21 @@ public class CompanyBankAccountService {
         }
     }
 
-    public CompanyBankAccount createBankAccount(CompanyBankAccount bankAccount) {
-        return bankAccountRepository.save(bankAccount);
+    public CompanyBankAccount createBankAccount(String companyId, CompanyBankAccountDto accountDto) {
+        Optional<Company> companyOptional = companyRepository.findById(companyId);
+        if(companyOptional.isEmpty()) {
+            throw new CompanyNotFoundException("There is no company with this ID.");
+        }
+
+        CompanyBankAccount companyBankAccount = new CompanyBankAccount();
+            companyBankAccount.setAccountNumber(accountDto.getAccountNumber());
+            companyBankAccount.setBankName(accountDto.getBankName());
+            companyBankAccount.setCurrency(accountDto.getCurrency());
+        bankAccountRepository.save(companyBankAccount);
+        companyOptional.get().getBankAccounts().add(companyBankAccount);
+        companyRepository.save(companyOptional.get());
+
+        return companyBankAccount;
     }
 
     public CompanyBankAccount updateBankAccount(CompanyBankAccountDto bankAccountDto) {
@@ -46,11 +65,18 @@ public class CompanyBankAccountService {
         }
     }
 
-    public void deleteBankAccount(String id) {
+    public void deleteBankAccount(String id, String companyId) {
+        Optional<Company> companyOptional = companyRepository.findById(companyId);
+        if(companyOptional.isEmpty()) {
+            throw new CompanyNotFoundException("There is no company with this ID.");
+        }
         Optional<CompanyBankAccount> bankAccount = bankAccountRepository.findById(id);
         if (bankAccount.isPresent()) {
+            companyOptional.get().getBankAccounts().remove(bankAccount.get());
+            companyRepository.save(companyOptional.get());
             bankAccountRepository.deleteById(id);
-        } else {
+        }
+        else {
             throw new BankAccountNotFoundException("Bank account not found with id: " + id);
         }
     }
